refactor: Remove unused code, simplify state management
